feat: add support for elb attributes
Add the new property attributes in load balancer entry. Also
refactor the module to isolate mappers
